renamed mt/st/ft sanitizer to something more generic
